IBX-5293: Replaced custom redirect success handler with listener
